rename new component
